requirements.txt
biopython==1.70
cycler==0.10.0
decorator==4.1.2
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
matplotlib==2.0.2
numpy==1.13.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
scipy==0.19.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7
